+++ freemocap_blender_addon/blender_ui/operators/_add_base_of_support.py
@@ -1,6 +1,6 @@
 import bpy
 
-from freemocap_blender_addon.blender_ui.ui_utilities import add_base_of_support
+from freemocap_blender_addon.core_functions.com_bos.add_bos import add_base_of_support
 from freemocap_blender_addon.core_functions.com_bos.add_com_vertical_projection import add_com_vertical_projection
 
 
@@ -14,9 +14,10 @@ class FREEMOCAP_OT_add_base_of_support(bpy.types.Operator):
         print("Adding Base of Support.......")
         ui_props = context.scene.freemocap_ui_properties
         # Check if the COM_Vertical_Projection object exists, if not create it
+        data_parent_empty = context.scene.freemocap_properties.data_parent_empty
+
         if bpy.data.objects.get("COM_Vertical_Projection") is None:
-            data_parent_object = context.scene.freemocap_properties.data_parent_empty
-            add_com_vertical_projection(data_parent_object=data_parent_object,
+            add_com_vertical_projection(data_parent_empty=data_parent_empty,
                                         neutral_color=ui_props.com_vertical_projection_neutral_color,
                                         in_bos_color=ui_props.com_vertical_projection_in_bos_color,
                                         out_bos_color=ui_props.com_vertical_projection_out_bos_color)
@@ -25,7 +26,8 @@ class FREEMOCAP_OT_add_base_of_support(bpy.types.Operator):
             ui_props.show_com_vertical_projection = True
 
         # Add Base of Support
-        add_base_of_support(z_threshold=ui_props.base_of_support_z_threshold,
+        add_base_of_support(data_parent_empty=data_parent_empty,
+                            z_threshold=ui_props.base_of_support_z_threshold,
                             point_of_contact_radius=ui_props.base_of_support_point_radius,
                             color=ui_props.base_of_support_color)
