+++ freemocap_blender_addon/blender_ui/operators/_add_com_vertical_projection.py
@@ -10,9 +10,9 @@ class FREEMOCAP_OT_add_com_vertical_projection(bpy.types.Operator):
 
     def execute(self, context):
         print("Adding COM Vertical Projection.......")
-        data_parent_object = context.scene.freemocap_properties.data_parent_empty
+        data_parent_empty = context.scene.freemocap_properties.data_parent_empty
         # Add COM Vertical Projection
-        add_com_vertical_projection(data_parent_object=data_parent_object,
+        add_com_vertical_projection(data_parent_empty=data_parent_empty,
                                     neutral_color=context.scene.freemocap_ui_properties.com_vertical_projection_neutral_color,
                                     in_bos_color=context.scene.freemocap_ui_properties.com_vertical_projection_in_bos_color,
                                     out_bos_color=context.scene.freemocap_ui_properties.com_vertical_projection_out_bos_color)
